Remove notification receiver
